--- a/src/utils/mod.rs
+++ b/src/utils/mod.rs
@@ -1,7 +1,6 @@
 //!  Utility functions for Rustup
 pub(crate) mod notifications;
 pub mod raw;
-pub(crate) mod toml_utils;
 pub(crate) mod units;
 #[allow(clippy::module_inception)]
 pub mod utils;