This saves a bunch of code (-190 lines) and is generally more idiomatic. It also saves about 8% of latency on proxy invocations, after (both best of 3 runs):

```
djc-2021 toml rustup $ RUSTUP_FORCE_ARG0=rustc hyperfine "./target/release/rustup-init --version"
Benchmark 1: ./target/release/rustup-init --version
  Time (mean ± σ):      32.6 ms ±   0.6 ms    [User: 23.3 ms, System: 5.5 ms]
  Range (min … max):    31.4 ms …  34.6 ms    86 runs
```

before:

```
djc-2021 master rustup $ RUSTUP_FORCE_ARG0=rustc hyperfine "./target/release/rustup-init --version"
Benchmark 1: ./target/release/rustup-init --version
  Time (mean ± σ):      35.3 ms ±   0.6 ms    [User: 25.6 ms, System: 5.8 ms]
  Range (min … max):    33.6 ms …  36.6 ms    79 runs
```